+++ kurtosis-module/impl/participant_network/cl/cl_client_launcher.go
@@ -20,6 +20,8 @@ type CLClientLauncher interface {
 		bootnodeContext *CLClientContext,
 		elClientContext *el.ELClientContext,
 		nodeKeystoreDirpaths *cl_validator_keystores.NodeTypeKeystoreDirpaths,
+		extraBeaconParams []string,
+		extraValidatorParams []string,
 	) (
 		resultClientCtx *CLClientContext,
 		resultErr error,