+++ kurtosis-module/impl/participant_network/cl/teku/teku_launcher.go
@@ -90,6 +90,8 @@ func (launcher *TekuCLClientLauncher) Launch(
 	bootnodeContext *cl.CLClientContext,
 	elClientContext *el.ELClientContext,
 	nodeKeystoreDirpaths *cl2.NodeTypeKeystoreDirpaths,
+	extraBeaconParams []string,
+	extraValidatorParams []string,
 ) (resultClientCtx *cl.CLClientContext, resultErr error) {
 
 	logLevel, err := module_io.GetClientLogLevelStrOrDefault(participantLogLevel, globalLogLevel, tekuLogLevels)
@@ -97,6 +99,7 @@ func (launcher *TekuCLClientLauncher) Launch(
 		return nil, stacktrace.Propagate(err, "An error occurred getting the client log level using participant log level '%v' and global log level '%v'", participantLogLevel, globalLogLevel)
 	}
 
+	extraParams := append(extraBeaconParams, extraValidatorParams...)
 	containerConfigSupplier := launcher.getContainerConfigSupplier(
 		image,
 		bootnodeContext,
@@ -104,6 +107,7 @@ func (launcher *TekuCLClientLauncher) Launch(
 		logLevel,
 		nodeKeystoreDirpaths.TekuKeysDirpath,
 		nodeKeystoreDirpaths.TekuSecretsDirpath,
+		extraParams,
 	)
 	serviceCtx, err := enclaveCtx.AddService(serviceId, containerConfigSupplier)
 	if err != nil {
@@ -148,6 +152,7 @@ func (launcher *TekuCLClientLauncher) getContainerConfigSupplier(
 	logLevel string,
 	validatorKeysDirpathOnModuleContainer string,
 	validatorSecretsDirpathOnModuleContainer string,
+	extraParams []string,
 ) func(string, *services.SharedPath) (*services.ContainerConfig, error) {
 	containerConfigSupplier := func(privateIpAddr string, sharedDir *services.SharedPath) (*services.ContainerConfig, error) {
 
@@ -232,6 +237,9 @@ func (launcher *TekuCLClientLauncher) getContainerConfigSupplier(
 		}
 		if bootnodeContext != nil {
 			cmdArgs = append(cmdArgs, "--p2p-discovery-bootnodes=" + bootnodeContext.GetENR())
+		}
+		if len(extraParams) > 0 {
+			cmdArgs = append(cmdArgs, extraParams...)
 		}
 		cmdStr := strings.Join(cmdArgs, " ")