+++ kurtosis-module/impl/participant_network/el/besu/besu_launcher.go
@@ -1,5 +1,4 @@
 package besu
-
 import (
 	"fmt"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/module_io"
@@ -70,14 +69,21 @@ func (launcher *BesuELClientLauncher) Launch(
 	participantLogLevel string,
 	globalLogLevel module_io.GlobalClientLogLevel,
 	bootnodeContext *el.ELClientContext,
+	extraParams []string,
 ) (resultClientCtx *el.ELClientContext, resultErr error) {
-
 	logLevel, err := module_io.GetClientLogLevelStrOrDefault(participantLogLevel, globalLogLevel, besuLogLevels)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting the client log level using participant log level '%v' and global log level '%v'", participantLogLevel, globalLogLevel)
 	}
 
-	containerConfigSupplier := launcher.getContainerConfigSupplier(image, launcher.networkId, bootnodeContext, logLevel)
+	containerConfigSupplier := launcher.getContainerConfigSupplier(
+		image,
+		launcher.networkId,
+		bootnodeContext,
+		logLevel,
+		extraParams,
+	)
+
 	serviceCtx, err := enclaveCtx.AddService(serviceId, containerConfigSupplier)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred launching the Besu EL client with service ID '%v'", serviceId)
@@ -116,6 +122,7 @@ func (launcher *BesuELClientLauncher) getContainerConfigSupplier(
 	networkId string,
 	bootnodeContext *el.ELClientContext, // NOTE: If this is empty, the node will be configured as a bootnode
 	logLevel string,
+	extraParams []string,
 ) func(string, *services.SharedPath) (*services.ContainerConfig, error) {
 	result := func(privateIpAddr string, sharedDir *services.SharedPath) (*services.ContainerConfig, error) {
 
@@ -152,6 +159,9 @@ func (launcher *BesuELClientLauncher) getContainerConfigSupplier(
 				launchNodeCmdArgs,
 				"--bootnodes=" + bootnodeContext.GetEnode(),
 			)
+		}
+		if len(extraParams) > 0 {
+			launchNodeCmdArgs = append(launchNodeCmdArgs, extraParams...)
 		}
 		launchNodeCmdStr := strings.Join(launchNodeCmdArgs, " ")